requirements.txt
paho-mqtt==1.5.0
PyYAML==5.3.1
influxdb==5.3.0
jsonpath-ng==1.5.2